+++ rest_framework_role_filters/role_filters.py
@@ -1,3 +1,7 @@
+import warnings
+
+__all__ = ['RoleFilter', 'RoleFilterGroup']
+
 class RoleFilter:
     role_id = None
 
@@ -5,16 +9,32 @@ class RoleFilter:
         selected_filter = getattr(self, filter_name)
         return selected_filter(*args, **kwargs)
 
-    def get_allowed_actions(self, request, view):
-        return ['create', 'list', 'retrieve', 'update', 'partial_update', 'destroy']
+    def get_allowed_actions(self, request, view, obj=None):
+        """Return container of allowed actions.
+
+        Defaults to list of all ViewSet actions.
+
+        By default RoleFilterMixin calls this method during `ViewSet.initialize` with `obj=None`,
+        and during `ViewSet.check_object_permissions` with `obj=something`.
+
+        Should NOT return `None`.
+
+        """
+        actions = getattr(view, 'action_map', {}) or {}
+        actions = list(actions.values())
+        actions = actions or ['create', 'list', 'retrieve', 'update', 'partial_update', 'destroy']
+        return actions
 
     def get_queryset(self, request, view, queryset):
+        """Return modified queryset for this role, or None if no modifications required."""
         return
 
     def get_serializer_class(self, request, view):
+        """Return modified serializer class for this role, or None if no modifications required."""
         return
 
     def get_serializer(self, request, view, serializer_class, *args, **kwargs):
+        """Return modified serializer instance for this role, or None if no modifications required."""
         return
 
 
@@ -31,8 +51,16 @@ class RoleFilterGroup:
             return
         return role_filter.trigger_filter(filter_name, *args, **kwargs)
 
-    def get_allowed_actions(self, role_id, request, view):
-        allowed_actions = self.trigger_filter('get_allowed_actions', role_id, request, view)
+    def get_allowed_actions(self, role_id, request, view, obj=None):
+        try:
+            allowed_actions = self.trigger_filter('get_allowed_actions', role_id, request, view, obj=obj)
+        except TypeError:
+            allowed_actions = self.trigger_filter('get_allowed_actions', role_id, request, view)
+            warnings.warn(
+                "RoleFilter.get_allowed_actions without support for `obj` argument is deprecated",
+                PendingDeprecationWarning
+            )
+
         if allowed_actions is None:
             return []
         return allowed_actions