+++ testproject/tests/testapp/test_views.py
@@ -1,3 +1,6 @@
+import json
+import warnings
+
 import pytest
 from django.urls import reverse
 from rest_framework import status
@@ -21,6 +24,49 @@ def test_get_allowed_actions_with_no_declared_role_inside_view(
     response = client_user_logged.delete(reverse('testapp:posts-detail', args=[post1.id]))
     assert response.status_code == status.HTTP_403_FORBIDDEN
     assert response.data == {'detail': 'action=destroy not allowed for role=no_post_access'}
+
+
+def test_get_allowed_actions_with_obj_level_role_inside_view(
+        post3, user_with_obj_level_role_in_view, client_user_logged):
+    url = reverse('testapp:posts-detail', args=[post3.id])
+    json_c_t = 'application/json'
+
+    response = client_user_logged.get(url)
+    assert response.status_code == status.HTTP_200_OK
+
+    response = client_user_logged.put(
+        url,
+        data={'title': 'new', 'body': 'new_body'},
+        content_type=json_c_t,
+    )
+    assert response.status_code == status.HTTP_403_FORBIDDEN
+    assert response.data == {'detail': 'action=update not allowed for role=obj_level_user'}
+
+    response = client_user_logged.patch(
+        url,
+        data=json.dumps({'title': 'new title'}),
+        content_type=json_c_t,
+    )
+    assert response.status_code == status.HTTP_200_OK
+
+    response = client_user_logged.get(reverse('testapp:posts-list'))
+    assert response.status_code == status.HTTP_200_OK
+
+
+def test_deprecated_get_allowed_actions(
+        post4, user_with_deprecated_role_in_view, client_user_logged):
+    with warnings.catch_warnings(record=True) as w:
+        warnings.simplefilter("always")
+        response = client_user_logged.delete(reverse('testapp:posts-detail', args=[post4.id]))
+        assert response.status_code == status.HTTP_403_FORBIDDEN
+        assert response.data == {'detail': 'action=destroy not allowed for role=deprecated_user'}
+
+        assert len(w) >= 1
+        assert issubclass(w[-1].category, PendingDeprecationWarning)
+        msg = str(w[-1].message)
+        assert "deprecated" in msg
+        assert "`obj` argument" in msg
+
 
 
 def test_get_queryset(